+++ lib/mobile/ntk_map_controller_mobile.dart
@@ -4,6 +4,8 @@ import '../interfaces/ntk_map_controller_interface.dart';
 import '../mobile/ntk_map_view_state_mobile.dart';
 import 'package:latlong2/latlong.dart';
 
+import '../modules/create_unique_uid.dart';
+
 ///NtkMapController for mobile versions of app
 class NtkMapController extends NtkMapControllerInterface {
   ///NtkMapController for mobile versions of app
@@ -34,7 +36,6 @@ class NtkMapController extends NtkMapControllerInterface {
         "_goToPointThenZoom(${point.latitude}, ${point.longitude}, $zoom)");
   }
 
-  ///For now not implemented on mobile, we create this method for new platform as soon as posible
   ///Create marker on point with title, in this marker you may configure a button and its callback
   ///on [point]
   ///with [title]
@@ -43,8 +44,28 @@ class NtkMapController extends NtkMapControllerInterface {
   @override
   addCustomMarker(
       LatLng point, String title, List<String> names, List<Function> acts) {
-    // TODO: implement addCustomMarker
-    throw UnimplementedError();
+    String buttIdsString = "[";
+
+    List<String> buttIds = [];
+    for (var act in acts) {
+      String id = createUniqueUid(count: 6, isNumberEnabled: false);
+
+      buttIds.add(id);
+      buttIdsString += '"$id", ';
+      //print(id);
+      super.markersAction[id] = act;
+    }
+
+    buttIdsString += "]";
+
+    String namesStr = "[";
+    for (var name in names) {
+      namesStr += '"$name", ';
+    }
+    namesStr += "]";
+
+    NtkMapViewState.controller.runJavaScript(
+        "_addMarkerCustom(${point.latitude}, ${point.longitude}, '$title', $namesStr, $buttIdsString)");
   }
 
   ///Create polyline on List of [points] (also clear all previous polyline)
@@ -57,9 +78,22 @@ class NtkMapController extends NtkMapControllerInterface {
         .runJavaScript("_createPolyline(${[pointsForMap]})");
   }
 
+  ///Update **[filter]** map
   @override
   applyNewFilter(MapFilter filter) {
-    // TODO: implement applyNewFilter
+    var list = filter.toParameterString();
+
+    NtkMapViewState.controller.runJavaScript(
+        '_updateFilter("${list[0]}", "${list[1]}", "${list[2]}", "${list[3]}", "${list[4]}", "${list[5]}", "${list[6]}", "${list[7]}", "${list[8]}")');
+  }
+
+  ///FOR NOW NOT IMPLEMENTED ON MOBILE
+  ///Update current position on map
+  ///This create a circle and marker with center in **[point]**
+  ///Radius of circle is **[accuracy]**
+  @override
+  updateCurrentPosition(LatLng point, double accuracy) {
+    // TODO: implement updateCurrentPosition
     throw UnimplementedError();
   }
 }