+++ lib/web/ntk_map_view_state_web.dart
@@ -92,8 +92,6 @@ class NtkMapViewState extends State<NtkMapViewInterface> {
         viewType: viewID,
         onPlatformViewCreated: (e) {
           widget.onCreateEnd!(widget.controller! as NtkMapController);
-
-
         },
       ),
     );